requirements.txt
adbutils==2.9.4
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
Appium-Python-Client==3.0.0
async-timeout==5.0.1
attrs==25.3.0
auto-py-to-exe==2.46.0
blinker==1.9.0
bottle==0.13.4
bottle-websocket==0.2.9
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
construct==2.10.70
decompyle3==3.9.2
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
discord.py==2.5.2
Eel==0.18.2
emulator-Thw==0.0.11
exceptiongroup==1.3.0
ExifRead==3.4.0
Faker==37.5.3
fastapi==0.115.12
Flask==3.1.1
frozenlist==1.7.0
future==1.0.0
geographiclib==2.0
geopy==2.4.1
gevent==25.5.1
gevent-websocket==0.10.1
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.74.0
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
MouseInfo==0.1.3
multidict==6.6.3
names==0.3.0
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
pefile==2023.2.7
piexif==1.1.3
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
py==1.11.0
py-cpuinfo==9.0.0
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.17
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.2
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.7
pyinstxtractor-ng==2025.1.6
PyMsgBox==1.0.9
pyotp==2.9.0
pyparsing==3.2.3
pyperclip==1.9.0
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
pyqt6-tools==6.4.2.3.3
PyQt6_sip==13.10.2
PyRect==0.2.0
Pyrogram==2.0.106
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-telegram-bot==20.7
pytweening==1.2.0
pywin32==311
pywin32-ctypes==0.2.3
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
requests==2.32.4
requests-oauthlib==2.0.0
retry==0.9.2
retry2==0.9.5
rich==14.1.0
rsa==4.9.1
selenium==4.34.2
shiboken6==6.9.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
spark-parser==1.8.9
SpeechRecognition==3.14.3
starlette==0.46.2
TgCrypto==1.2.5
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uiautodev==0.9.0
uiautomator2==3.4.0
uncompyle6==3.9.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
wdapy==0.2.4
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.2.0
xdis==6.1.3
yarl==1.20.1
yt-dlp==2025.7.21
zope.event==5.1.1
zope.interface==7.2
setuptools==79.0.1
dc-license-system==0.0.4
dc-thw-emulator==0.0.1
dc-u2-driver==0.0.1